31/7/2022 Finalizacion de Mantenimientos

+++ GymPro.Capa.Datos/Implementaciones/ExpedienteUsuarioDAL.cs
@@ -192,7 +192,7 @@ namespace GymPro.Capa.Datos.Implementaciones
                             Fecha = DateTime.Parse(reader["Fecha"].ToString()),
                             IdentificacionUsuario = reader["IdentificacionUsuario"].ToString(),
                             Peso = double.Parse(reader["Peso"].ToString()),
-                            Altura = double.Parse(reader["Altura"].ToString()),
+                            Altura = int.Parse(reader["Altura"].ToString()),
                             IMC = double.Parse(reader["IMC"].ToString()),
                             MetabolismoBasal = double.Parse(reader["MetabolismoBasal"].ToString())
                         };
@@ -249,7 +249,7 @@ namespace GymPro.Capa.Datos.Implementaciones
                             Fecha = DateTime.Parse(dr["Fecha"].ToString()),
                             IdentificacionUsuario = dr["IdentificacionUsuario"].ToString(),
                             Peso = double.Parse(dr["Peso"].ToString()),
-                            Altura = double.Parse(dr["Altura"].ToString()),
+                            Altura = int.Parse(dr["Altura"].ToString()),
                             IMC = double.Parse(dr["IMC"].ToString()),
                             MetabolismoBasal = double.Parse(dr["MetabolismoBasal"].ToString())
                         };
@@ -306,7 +306,7 @@ namespace GymPro.Capa.Datos.Implementaciones
                             Fecha = DateTime.Parse(dr["Fecha"].ToString()),
                             IdentificacionUsuario = dr["IdentificacionUsuario"].ToString(),
                             Peso = double.Parse(dr["Peso"].ToString()),
-                            Altura = double.Parse(dr["Altura"].ToString()),
+                            Altura = int.Parse(dr["Altura"].ToString()),
                             IMC = double.Parse(dr["IMC"].ToString()),
                             MetabolismoBasal = double.Parse(dr["MetabolismoBasal"].ToString())
                         };
