+++ GymPro.Capa.Entidades/Implementaciones/ExpedienteUsuario.cs
@@ -12,7 +12,7 @@ namespace GymPro.Capa.Entidades.Implementaciones
         public DateTime Fecha { get; set; }
         public string IdentificacionUsuario { get; set; }
         public double Peso { get; set; }
-        public double Altura { get; set; }
+        public int Altura { get; set; }
         public double IMC { get; set; }
         public double MetabolismoBasal { get; set; }
         #endregion
